+++ src/types/amfobject.hpp
@@ -101,7 +101,7 @@ public:
 		}
 
 		// U29-traits = 0b0011 = 0x03
-		int traitMarker = traits.attributes.size() << 4 | 0x03;
+		size_t traitMarker = traits.attributes.size() << 4 | 0x03;
 		// dynamic marker = 0b1000 = 0x08
 		if (traits.dynamic) traitMarker |= 0x08;
 
@@ -208,7 +208,7 @@ public:
 		}
 
 		AmfObject ret(traits);
-		int contextIndex = ctx.addObject<AmfObject>(ret);
+		size_t contextIndex = ctx.addObject<AmfObject>(ret);
 
 		if (traits.externalizable) {
 			ret = Deserializer::externalDeserializers.at(traits.className)(it, end, ctx);
